requirements.txt
faiss-cpu==1.8.0
sentence-transformers==3.0.1
rank-bm25==0.2.2
pdfplumber==0.11.2
pymupdf==1.24.7
google-generativeai==0.7.2
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
numpy==1.26.4
tqdm==4.66.4
groq>=0.9.0
